+++ Code/ItemEdits/ATG.cs
@@ -8,7 +8,7 @@ using MonoDetour;
 using MonoDetour.HookGen;
 using MonoDetour.Cil;
 using RoR2;
-using LordsItemEdits.MultiItemEdits;
+using RoR2.Orbs;
 namespace LordsItemEdits.ItemEdits;
 
 
@@ -25,6 +25,7 @@ internal static class ATG
 
         Mdh.RoR2.GlobalEventManager.ProcessHitEnemy.ILHook(FireOrbIfItsFair);
     }
+
 
 
     private static void FireOrbIfItsFair(ILManipulationInfo info)
@@ -67,8 +68,67 @@ internal static class ATG
     private static bool FireMissileOrbReturnFairness(CharacterBody attackerBody, float missileDamage, DamageInfo damageInfo, GameObject victim)
     {
         // orb won't fire if it's unfair
-        Missiles.FireMissileOrb(attackerBody, missileDamage, damageInfo, victim);
+        FireMissileOrb(attackerBody, missileDamage, damageInfo, victim);
         // we need to return if it's fair or not (aka if we should skip the firemissile line or not)
         return attackerBody.teamComponent.teamIndex == TeamIndex.Player;
+    }
+
+
+
+    internal static void FireMissileOrb(CharacterBody attackerBody, float missileDamage, DamageInfo damageInfo, GameObject victim)
+    {
+        if (victim == null || attackerBody.teamComponent.teamIndex != TeamIndex.Player)
+        {
+            return;
+        }
+        CharacterBody victimBody = victim.GetComponent<CharacterBody>();
+        FireMissileOrb(attackerBody, missileDamage, damageInfo, victimBody, true);
+    }
+
+    internal static void FireMissileOrb(CharacterBody attackerBody, float missileDamage, DamageInfo damageInfo, CharacterBody victimBody, bool addMissileProc)
+    {
+        if (victimBody == null || attackerBody == null || attackerBody.inventory == null || attackerBody.teamComponent.teamIndex != TeamIndex.Player)
+        {
+            return;
+        }
+
+
+        MicroMissileOrb missileOrb = new()
+        {
+            origin = attackerBody.aimOrigin,
+            damageValue = missileDamage,
+            isCrit = damageInfo.crit,
+            teamIndex = attackerBody.teamComponent.teamIndex,
+            attacker = attackerBody.gameObject,
+            procChainMask = damageInfo.procChainMask,
+            procCoefficient = 1f,
+            damageColorIndex = DamageColorIndex.Item,
+            target = victimBody.mainHurtBox
+        };
+        if (addMissileProc)
+        {
+            missileOrb.procChainMask.AddProc(ProcType.Missile);
+        }
+
+
+        if (
+            ConfigOptions.PocketICBM.EnableEdit.Value &&
+            (ConfigOptions.PocketICBM.ChangeATGEffect.Value && addMissileProc || ConfigOptions.PocketICBM.ChangeArmedBackpackEffect.Value && !addMissileProc)
+        )
+        {
+            // already hooked GetMoreMissileDamageMultiplier to do the edited version's damage so it's fine to use here
+            missileOrb.damageValue *= MissileUtils.GetMoreMissileDamageMultiplier(attackerBody.inventory.GetItemCountEffective(DLC1Content.Items.MissileVoid));
+        }
+        else
+        {
+            OrbManager.instance.AddOrb(missileOrb);
+            OrbManager.instance.AddOrb(missileOrb);
+            // gotta be authentic with the missile spam experience lmao
+            Util.PlaySound("Play_item_proc_missile_fire", attackerBody.gameObject);
+            Util.PlaySound("Play_item_proc_missile_fire", attackerBody.gameObject);
+        }
+        OrbManager.instance.AddOrb(missileOrb);
+        // the orb doesn't play a sound on fire and editing the assets isn't working so
+        Util.PlaySound("Play_item_proc_missile_fire", attackerBody.gameObject);
     }
 }